refactor: common/entity 로 분리
