+++ server-toss/src/common/entitiy/base.entity.ts
