+++ server-toss/src/modules/ad/ad-view.entity.ts
@@ -4,7 +4,7 @@ import {
   ManyToOne,
   PrimaryKey,
 } from "@mikro-orm/decorators/legacy";
-import { BaseEntity } from "src/common/base.entity";
+import { BaseEntity } from "src/common/entitiy/base.entity";
 import { User } from "../user/user.entity";
 import type { Rel } from "@mikro-orm/core";
