+++ server-toss/src/modules/drawing/drawing.entity.ts
@@ -5,7 +5,7 @@ import {
   PrimaryKey,
   Property,
 } from "@mikro-orm/decorators/legacy";
-import { BaseEntity } from "src/common/base.entity";
+import { BaseEntity } from "src/common/entitiy/base.entity";
 import { Prompt } from "src/modules/prompt/prompt.entity";
 import { User } from "src/modules/user/user.entity";
