+++ server-toss/src/modules/point/point-log.entity.ts
@@ -6,7 +6,7 @@ import {
   PrimaryKey,
   Property,
 } from "@mikro-orm/decorators/legacy";
-import { BaseEntity } from "src/common/base.entity";
+import { BaseEntity } from "src/common/entitiy/base.entity";
 import { User } from "src/modules/user/user.entity";
 
 @Entity({ tableName: "point_logs" })
